feat: 新增權限Type列舉

+++ backend2/src/main/java/com/cmoney/backend2/authorization/service/AuthorizationService.kt
@@ -1,6 +1,6 @@
 package com.cmoney.backend2.authorization.service
 
-import com.cmoney.backend2.authorization.service.api.ExpiredTime
+import com.cmoney.backend2.authorization.service.api.getexpiredtime.ExpiredTime
 import com.cmoney.backend2.base.model.calladapter.RecordApi
 import retrofit2.Response
 import retrofit2.http.GET
@@ -14,7 +14,7 @@ interface AuthorizationService {
     /**
      *取得會員某種權限類型的基底功能權限到期日(到期日是沒加時區的Utc時戳)
      *
-     * @param type 權限類型 (手機請使用 MobilePaid)
+     * @param type 權限類型
      * @param subjectId appId
      *
      */
@@ -23,6 +23,6 @@ interface AuthorizationService {
     suspend fun getExpiredTime(
         @Header("Authorization") authorization: String,
         @Path("type") type: String,
-        @Path("subjectId") subjectId: Int
+        @Path("subjectId") subjectId: Long
     ): Response<ExpiredTime>
 }
